+ Rewritten Auth code (Multi-level/Multi-class).
Possible a little overcomplex for what it needs to do, but it works
okay for now.  Should allow users to be logged in to a website
several times (one instance for user, one for administrator, but both
in the same session).

Idea being that your admin user is kept 100% separate and can not
be used to fiddle with the front-end functionality (but this could
be customised quite easily).

May need to look into session timeouts, since it will kill both
sessions at the same time.  Possibility of splitting into two
separate sessions?  We'll see...